Move Carla and RobustBench stuff into examples.
