+++ examples/carla_overhead_object_detection/configs/datamodule/armory_carla_over_objdet.yaml
