+++ examples/carla_overhead_object_detection/requirements.txt
@@ -1,0 +1,1 @@
+mart @ git+https://github.com/IntelLabs/MART.git
