+++ examples/carla_overhead_object_detection/tests/helpers
@@ -1,0 +1,1 @@
+../../../tests/helpers/
