+++ examples/carla_overhead_object_detection/tests/test_configs.py
@@ -1,0 +1,1 @@
+../../../tests/test_configs.py
